fix(user-api-keys): align frontend checks
